Fix bug
Allow it to default to 0. 

Use case is when I want to get all user posts. Without having to guess how many posts user has. There is a check elsewhere that fails if count is greater then user post count. 

In that case, if count is greater than user count, just return posts.